requirements.txt
opencv-python
onnxruntime
picamera2
flask
python-dotenv
requests
Werkzeug
